pyproject.toml
[tool.poetry]
name = "ecole-factory"
version = "0.1.0"
description = "3-AI company demo"
authors = ["OpenAI Assistant <assistant@example.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = {extras=["standard"], version="^0.29"}
sqlalchemy = "^2.0"
asyncpg = "^0.29"
aiosqlite = "^0.19"
alembic = "^1.13"
pydantic = "^2.5"
httpx = "^0.27"
pydantic-settings = "^2.4"
pyyaml = "^6.0"
jinja2 = "^3.1"
apscheduler = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
httpx = "^0.27"
black = "^24.3"
ruff = "^0.4"
mypy = "^1.9"
hypothesis = "^6.103"
freezegun = "^1.4"
pytest-cov = "^5.0"
mutmut = {version = "^2.4", optional = true}
pytest-playwright = {version = "^0.4", optional = true}

[tool.pytest.ini_options]
addopts = "--asyncio-mode=auto --cov=src/app/math --cov-report=term-missing --cov-fail-under=90"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
